+++ app/Http/Controllers/ThreadController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Auth;
 use App\Thread;
 use App\Reply;
 use App\User;
@@ -27,42 +28,38 @@ class ThreadController extends Controller
      */
     public function create(Request $request)
     {
+        // Om inte inloggad eller bannad
+        if(!Auth::check() || Auth::user()-> banned == 1)
+            return view('/forum');
 
-        // Skapa ny tråd
         $thread = new Thread;
         $thread->title = $request->title;
         $thread->body = $request->body;
         $thread->forum_id = $request->forumID;
-        $thread->user_id = 4;       // SKA komma från den inloggade användaren
+        $thread->user_id = Auth::user()->id;   
+        
+        $forum = Forum::find($request->forumID); 
+        $thread->rights = $forum->rights;
         $thread->save();
-
-        // Hämta användaren (för att öka antalet posts)
-        // $user = User::where('id', 4); // Ska komma från inloggade användaren
-        // $user->posts = $user->posts +1; 
-
-        // Hämta forumet som tråden tillhör för att öka antalet posts 
-        $forum = Forum::find($request->forumID);
+        
         $forum->num_threads = $forum->num_threads + 1;
         $forum->latest_thread = $thread->id;
-        $forum->save();
+        $forum->save();        
 
-        // Hämta senaste tråden i forumet man är i (den som precis skapades)
-        // Så att man kan returnera datan från den för att visa datan (har inte tillgång till id)
-        // annars
+        $user = Auth::user();
+        $user->num_posts = $user->num_posts + 1;
+        $user->save(); 
+
+        
         $thread = Thread::where('forum_id', $request->forumID)->latest()->first();
-
-        // dd($thread);
-        // Hämta replies
         $replies = Reply::where('thread_id', $thread->id)->get();
-        $users = User::all(); // borde bara hämta vissa ??
-
+        
 
         // Använder return redirect istället för return view för att 
         return redirect('/forum/'.$thread->forum_id.'/thread/'.$thread->id.'')
             ->with('from', 'thread')
             ->with('thread', $thread)
-            ->with('replies', $replies)
-            ->with('users', $users);                          
+            ->with('replies', $replies);                          
     }
 
     /**
@@ -85,20 +82,36 @@ class ThreadController extends Controller
      */
     public function show(Thread $thread, $forumID, $threadID)
     {
-        //
+        // Uppdatera antalet visningar
         $thread = Thread::find($threadID);
         $thread->num_views = $thread->num_views + 1;
         $thread->save();
 
-
-
         $replies = Reply::where('thread_id', $threadID)->get();
-        $users = User::all(); // borde bara hämta vissa ??
 
-        return view('forum')->with('from', 'thread')
-                            ->with('thread', $thread)
-                            ->with('replies', $replies)
-                            ->with('users', $users);        
+        // Skapa array av användar id:n
+        $repliers_id = [];
+        foreach ($replies as $reply){
+            array_push($repliers_id, $reply->user_id);
+        }
+
+        // hämta användarna som har skrivit i tråden
+        $repliers = User::all()->intersect(User::whereIn('id', $repliers_id)->get());
+
+        // Hämta forumet, så man kan verifera om användaren har rätt till forumet
+        $forum = Forum::find($forumID);
+
+        if(Auth::user() == null && $forum->rights == 'user' || Auth::user()->role <= $forum->rights){
+            return view('forum')->with('from', 'thread')
+                                ->with('thread', $thread)
+                                ->with('replies', $replies)
+                                ->with('repliers', $repliers)
+                                ->with('orignal_poster', User::find($thread->user_id));            
+        }else{
+            return redirect('/forum');
+        }
+
+
     }
 
     /**
@@ -130,9 +143,30 @@ class ThreadController extends Controller
      * @param  \App\Models\Thread  $thread
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Thread $thread, $threadID)
+    public function destroy(Thread $thread, $forumID, $threadID)
     {
         //
+        if(!Auth::check() || Auth::user()->role != 'admin'){
+            return redirect('/forum');
+        }
+        
+        $forum = Forum::find($forumID);
+        $forum->num_threads = $forum->num_threads - 1;
+        $forum->save();
         $thread->destroy($threadID);
+        $threads = Thread::where('forum_id', $forum->id)->get();
+
+
+        $latest_replies = collect();    
+        foreach($threads as $thread){
+            $reply = Reply::where('thread_id', $thread->id)->orderBy('updated_at','DESC')->get()->first();
+            if($reply != null)
+                $latest_replies->push($reply);        
+        }
+
+        return view('forum')->with('from', 'forum')
+                            ->with('forum', $forum)
+                            ->with('threads', $threads)
+                            ->with('latest_replies', $latest_replies);
     }
 }
